pyproject.toml
[project]
name = "llm-tool-calling"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "groq>=0.32.0",
    "langchain-core>=0.3.76",
    "langchain-google-genai>=2.0.10",
    "python-dotenv>=1.1.1",
]
